% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{plot.JoF}
\alias{plot.JoF}
\title{plot progress of judgment of frequency simulation}
\usage{
\method{plot}{JoF}(x, type = "l", ...)
}
\arguments{
\item{x}{output of JoF simulation}

\item{type}{"l" for lines, "p" for points, "b" for both}

\item{...}{further arguments}
}
\value{
Displays the judgment of frequency as proportion of all
inputs
}
\description{
plot progress of judgment of frequency simulation
}
