% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.JoF}
\alias{print.JoF}
\title{Output of judgment of frequecny simulation}
\usage{
\method{print}{JoF}(x, ...)
}
\arguments{
\item{x}{output of JoF simulation}

\item{...}{further arguments}
}
\value{
Displays the judgment of frequency as proportion of all
inputs
}
\description{
Output of judgment of frequecny simulation
}
